requirements.txt
python==3.10.15
faiss-cpu==1.7.4
langchain==0.1.13
langchain-openai==0.0.7
numpy==1.26.4
pandas==2.2.1
PyPDF2==3.0.1
python-docx==1.0.1
tqdm==4.66.2
openai==1.12.0